fix(events): factor out abstract classes
