+++ packit_service/worker/events/pagure/abstract.py
@@ -7,12 +7,12 @@ from typing import Optional
 from packit.config import PackageConfig
 
 from packit_service.config import PackageConfigGetter
-from packit_service.worker.events.event import AbstractForgeIndependentEvent
+from packit_service.worker.events.abstract import ForgeIndependent
 
 logger = getLogger(__name__)
 
 
-class PagureEvent(AbstractForgeIndependentEvent):
+class PagureEvent(ForgeIndependent):
     def __init__(self, project_url: str, pr_id: Optional[int] = None, **kwargs):
         super().__init__(pr_id=pr_id)
         self.project_url: str = project_url
